+++ setup/foundry.sh
@@ -86,8 +86,14 @@ find /foundrydata -type f -exec chmod 664 {} +
 cp /aws-foundry-ssl/setup/foundry/foundry.service /etc/systemd/system/foundry.service
 chmod 644 /etc/systemd/system/foundry.service
 
+# Set up health check to restart Foundry if unresponsive (addresses V11+ update behaviour)
+cp /aws-foundry-ssl/setup/foundry/foundry_healthcheck.sh /foundrycron/foundry_healthcheck.sh
+cp /aws-foundry-ssl/setup/foundry/foundry_healthcheck.service /etc/systemd/system/foundry_healthcheck.service
+cp /aws-foundry-ssl/setup/foundry/foundry_healthcheck.timer /etc/systemd/system/foundry_healthcheck.timer
+
 systemctl daemon-reload
 systemctl enable --now foundry
+systemctl enable --now foundry_healthcheck.timer
 
 # Configure foundry aws json file
 F_DIR='/foundrydata/Config/'